// src/app/models/product.interface.ts
export interface ProductI{
    titlePost: string;
    telPost: number;
    precioPost: number;
    dirPost:string;
    id?: string;
    categoriPost: string;
    imagePost1?: any;
    imagePost2?: any;
    imagePost3?: any;
    fileRef1?: string;
    fileRef2?: string;
    fileRef3?: string;
    namePost: string;
    tagsPost: string;
    permisosPost:string;
}